+++ src/main/java/com/bidnamu/bidnamubackend/auth/dto/request/CreateAccessTokenRequest.java
@@ -1,4 +1,4 @@
-package com.bidnamu.bidnamubackend.auth.dto;
+package com.bidnamu.bidnamubackend.auth.dto.request;
 
 
 public record CreateAccessTokenRequest(String refreshToken) {
