+++ src/main/java/com/bidnamu/bidnamubackend/auth/dto/response/CreateAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package com.bidnamu.bidnamubackend.auth.dto;
+package com.bidnamu.bidnamubackend.auth.dto.response;
 
 
 public record CreateAccessTokenResponse(String accessToken) {
